Updating Jeep-Sales video application where we added code to handle an order request from start to finish, including saving the order request to the database and testing to ensure the order was saved properly and generated a 201 status code.

+++ src/main/java/com/promineotech/jeep/dao/DefaultJeepSalesDao.java
@@ -57,7 +57,7 @@ public class DefaultJeepSalesDao implements JeepSalesDao {
         @Override
         public Jeep mapRow(ResultSet rs, int rowNum) throws SQLException {
           return Jeep.builder()
-            .modelPK(rs.getInt("model_pk"))
+            .modelPK(rs.getLong("model_pk"))
             .modelId(JeepModel.valueOf(rs.getString("model_id")))
             .trimLevel(rs.getString("trim_level"))
             .numDoors(rs.getInt("num_doors"))
